BUG #11382. Fixed errors when client-server with 32-64 bit.
There was issue with sizes of message passed around when client-server had
different sizes of unsigned long. This was causing the connection problems
reported in the bug. fixed that.

+++ Servers/Filters/vtkSpreadSheetView.cxx
@@ -312,7 +312,7 @@ void vtkSpreadSheetView::Update()
   this->SomethingUpdated = false;
   this->Superclass::Update();
 
-  unsigned long num_rows = 0;
+  unsigned int num_rows = 0;
 
   // From the active representation obtain the data/selection producers that
   // need to be streamed to the client.
